Inline RLSTimestampCorrector into athena.py
timestamps.py was a "shared" module with a single consumer (athena.py) —
legacy Muse kept its own copy of the same RLS math, so nothing was actually
shared. Move the class into athena.py (its only user) and drop the file; if
Muse later adopts it, extract then.
